# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentación de la API del sistema de gestión",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/especializaciones": {
            "get": {
                "tags": [
                    "Especializaciones"
                ],
                "summary": "Obtener todas las especializaciones",
                "operationId": "f8951077b6ad0b382c7bad0d89bb7998",
                "responses": {
                    "200": {
                        "description": "Lista de especializaciones"
                    },
                    "500": {
                        "description": "Error al obtener especializaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Especializaciones"
                ],
                "summary": "Crear una nueva especialización",
                "operationId": "3758bd4959878d8bfec7182ab02c47ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mecánica Automotriz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Especialización creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/especializaciones/{id}": {
            "get": {
                "tags": [
                    "Especializaciones"
                ],
                "summary": "Obtener una especialización específica",
                "operationId": "3b2330f62eddb2b03d7171d9e00508fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la especialización",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialización encontrada"
                    },
                    "404": {
                        "description": "Especialización no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Especializaciones"
                ],
                "summary": "Actualizar una especialización existente",
                "operationId": "4aaf89ef11f9f833b7936a4bad7ac4f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la especialización a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Electrónica Industrial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Especialización actualizada exitosamente"
                    },
                    "404": {
                        "description": "Especialización no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Especializaciones"
                ],
                "summary": "Eliminar una especialización",
                "operationId": "48ae5389bd945b6ee56cf9e6ae73d228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la especialización a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialización eliminada exitosamente"
                    },
                    "404": {
                        "description": "Especialización no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}